third_party/WebKit/LayoutTests/bluetooth/requestDevice.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

test(function(t) { assert_exists(window, "testRunner"); t.done(); },
     "window.testRunner is required for these tests.");

async_test(function(t) {
  testRunner.setBluetoothMockDataSet('RejectRequestDevice_NotFoundError');
  navigator.bluetooth.requestDevice()
    .then(
      t.step_func(function() { assert_unreached('Mock was set to reject.'); }),
      t.step_func(function(e) {
        assert_equals(e.name, 'NotFoundError');
        t.done();
  }));
}, 'Reject with NotFoundError.');

async_test(function(t) {
  testRunner.setBluetoothMockDataSet('RejectRequestDevice_SecurityError');
  navigator.bluetooth.requestDevice()
    .then(
      t.step_func(function() { assert_unreached('Mock was set to reject.'); }),
      t.step_func(function(e) {
        assert_equals(e.name, 'SecurityError');
        t.done();
  }));
}, 'Reject with SecurityError.');

async_test(function(t) {
  testRunner.setBluetoothMockDataSet('ResolveRequestDevice_Empty');
  navigator.bluetooth.requestDevice()
    .then(
      t.step_func(function() { t.done() }),
      t.step_func(function() { assert_unreached('Mock was set to resolve.'); }));
}, 'Mock will resolve.');

</script>
